+++ src/lib.rs
@@ -12,4 +12,4 @@ mod calculate_amounts;
 pub use ledger_parser::{
     Amount, Commodity, CommodityPosition, CommodityPrice, Reality, TransactionStatus,
 };
-pub use simplified_ledger::{Ledger, Posting, SimplificationError, Transaction};
+pub use simplified_ledger::{Error, Ledger, Posting, Transaction};