+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+- Add FromStr impl for simplified_ledger::Ledger
+- Apply parsed Transaction metadata to Postings in simplified_ledger::Ledger
+- Rename SimplificationError to Error and wrap ledger_parser::ParseError into it
+- Remove zero-balance accounts from Balance when adding transactions
+
 ## [0.5.0] - 2024-03-05
 
 - Update ledger-parser to 6